Mock html elements
